# contract/sdk-generator/_cut/admin-web.openapi.yaml
openapi: 3.0.3
info:
  title: 上医智养堂 · 三端接口契约（调元云）
  version: api-contract-v1.0.0
  description: "三端（客户小程序 / 调理师·经络师 APP / 管理员 Web）一次冻结的接口契约。\n\n版本号规则（语义化，与 info.version 同值）：\n  - MAJOR（破坏级）：字段删除 / 改名 / 类型变更；可见性档位变更；枚举值语义变更；错误码语义变更；\n    涉可见性档位的 MAJOR 变更须触发「变更业务裁定」（不得按\"体验优化\"处理）。\n  - MINOR（兼容级）：新增可选请求字段 / 新增接口 / 新增响应字段。\n  - PATCH（文档级）：注释 / 示例 / 措辞修正，不动字段与语义。\n\n\U0001F6D1 措辞纪律：client_end=mp 可见字段不得含\"退款\"\"诊断\"；同步时间精确到日。\n"
  contact:
    name: 后端主程（技术单签）· 产品共签（可见性 / 措辞档位）
  license:
    name: Proprietary
servers:
- url: /api/v1
  description: Base Path（契约 §2.0 全局约定）
...
paths:
...
  /plans/{id}:
    get:
      operationId: getPlan
      tags:
      - D-服务与履约
      summary: D5-b 方案查阅
      x-contract-row: D5
      x-callable-roles:
      - client
      - therapist
      - meridian
      - admin
      x-client-forbidden: false
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 方案详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /plans/{id}/reviews:
    post:
      operationId: reviewPlan
      tags:
      - D-服务与履约
      summary: D5-c 方案审核（退回必填 reason）
      x-contract-row: D5
      x-callable-roles:
      - meridian
      - admin
      x-client-forbidden: false
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 审核结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultEnvelope'
        '409':
          $ref: '#/components/responses/VersionConflict'
  /device-dispatches:
    post:
      operationId: createDeviceDispatch
      tags:
      - D-服务与履约
      summary: D6 设备参数下发（需方案已审核）
      description: '🛑 本接口操作的是【门店级调理设备 device】（下行、可追责），

        与【客户级手环 band】是两本台账、不得合并。

        '
      x-contract-row: D6
      x-callable-roles:
      - therapist
      - meridian
      - admin
      x-client-forbidden: false
      responses:
        '200':
          description: 下发结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultEnvelope'
        '403':
          $ref: '#/components/responses/GateMissing'
...
components:
...
  responses:
...
    GateMissing:
      description: 403 GATE_MISSING（前置门禁缺失；data.missing_items[] 给出缺失项名称，不得模糊报错）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
...
    NotFound:
      description: 404 NOT_FOUND（仅限本租户内确实不存在）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
    VersionConflict:
      description: 409 VERSION_CONFLICT（版本不可覆盖）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultEnvelope'
...
  schemas:
    ResultEnvelope:
      type: object
      required:
      - code
      - message
      - trace_id
      properties:
        code:
          type: integer
          description: 0 = 成功；非 0 见 x-error-codes
        message:
          type: string
          description: 面向开发者，不得直接渲染给客户（客户端有自己的 code → copy 映射）
        data:
          type: object
          nullable: true
          description: code != 0 时为空
        trace_id:
          type: string
          x-visible-to:
          - client
          - therapist
          - meridian
          - admin
      x-visibility-note: 信封字段对全部角色可见
...